pyproject.toml
[tool.poetry]
name = "Game Recommender"
version = "0.1.0"
description = ""
authors = ["freakynoblegas <nickquinn3131@gmail.com>", "Wolfinbarger <xzaiver.wolfinbarger@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlmodel = "^0.0.18"
fastapi = "^0.111.0"
python-dotenv = "^1.0.1"
alembic = "^1.13.1"
psycopg2 = "^2.9.9"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"